backend/requirements.txt
fastapi
uvicorn
aiofiles
python-multipart
gitpython
requests